requirements.txt
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
datasets>=2.16.0
safetensors>=0.4.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Optional but recommended
matplotlib>=3.7.0
wandb>=0.16.0
bitsandbytes>=0.41.0
xformers>=0.0.22
